LibURL: Keep the blob URL entry in its own header
